data_project/requirements.txt
psycopg2-binary
Faker
apache-airflow